backend/requirements.txt
Flask-SocketIO>=5.5.*
pydantic>=2.11.2
SQLAlchemy==2.0.*
python-ulid==3.0.0
python-dotenv==1.1.0
fastapi[standard]
pyjwt[crypto]
passlib[bcrypt]
loguru
